# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Condition": {
        "additionalProperties": false,
        "properties": {
          "key": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Key",
            "type": "string"
          },
          "operator": {
            "enum": [
              "equals",
              "semver"
            ],
            "title": "Operator",
            "type": "string"
          },
          "value": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "key",
          "operator",
          "value"
        ],
        "title": "Condition",
        "type": "object"
      },
      "FileDescriptor": {
        "additionalProperties": false,
        "properties": {
          "media_type": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Media Type",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "size": {
            "maximum": 20971520.0,
            "minimum": 0.0,
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "sha256",
          "size",
          "media_type"
        ],
        "title": "FileDescriptor",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ReleaseRecord": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "namespace": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Namespace",
            "type": "string"
          },
          "state": {
            "enum": [
              "preparing",
              "published",
              "yanked",
              "blocked"
            ],
            "title": "State",
            "type": "string"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/TargetRecord"
            },
            "title": "Targets",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "namespace",
          "name",
          "version",
          "state",
          "targets"
        ],
        "title": "ReleaseRecord",
        "type": "object"
      },
      "TargetAssociation": {
        "additionalProperties": false,
        "properties": {
          "build_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Id"
          },
          "manifest": {
            "$ref": "#/components/schemas/TargetManifest"
          },
          "source_repository": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Source Repository",
            "type": "string"
          },
          "source_revision": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Source Revision",
            "type": "string"
          }
        },
        "required": [
          "manifest",
          "source_repository",
          "source_revision"
        ],
        "title": "TargetAssociation",
        "type": "object"
      },
      "TargetManifest": {
        "additionalProperties": false,
        "properties": {
          "artifact_format": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Artifact Format",
            "type": "string"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/Condition"
            },
            "title": "Conditions",
            "type": "array"
          },
          "entrypoint": {
            "title": "Entrypoint",
            "type": "string"
          },
          "files": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FileDescriptor"
            },
            "maxProperties": 2048,
            "minProperties": 1,
            "title": "Files",
            "type": "object"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "artifact_format",
          "entrypoint",
          "conditions",
          "files"
        ],
        "title": "TargetManifest",
        "type": "object"
      },
      "TargetRecord": {
        "additionalProperties": false,
        "properties": {
          "artifact_format": {
            "title": "Artifact Format",
            "type": "string"
          },
          "build_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Id"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/Condition"
            },
            "title": "Conditions",
            "type": "array"
          },
          "entrypoint": {
            "title": "Entrypoint",
            "type": "string"
          },
          "source_repository": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Repository"
          },
          "source_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Revision"
          },
          "target_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Target Digest",
            "type": "string"
          },
          "target_key": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Target Key",
            "type": "string"
          }
        },
        "required": [
          "target_key",
          "target_digest",
          "artifact_format",
          "entrypoint",
          "conditions"
        ],
        "title": "TargetRecord",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Public metadata/artifact reads and scoped Extension target publication.",
    "title": "InKCre Extension Registry",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/livez": {
      "get": {
        "operationId": "livez_livez_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Livez Livez Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Livez"
      }
    },
    "/v1/artifacts/{target_digest}/files/{relative_path}": {
      "get": {
        "operationId": "get_artifact_file",
        "parameters": [
          {
            "in": "path",
            "name": "target_digest",
            "required": true,
            "schema": {
              "pattern": "^sha256:[0-9a-f]{64}$",
              "type": "string"
            }
          },
          {
            "description": "Normalized POSIX path, including nested segments.",
            "in": "path",
            "name": "relative_path",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Immutable target file bytes."
          },
          "404": {
            "description": "Target or file not found."
          },
          "451": {
            "description": "The associated release is operator-blocked."
          }
        },
        "summary": "Read one digest-addressed target file",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/v1/artifacts/{target_digest}/manifest": {
      "get": {
        "operationId": "get_target_manifest_v1_artifacts__target_digest__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "target_digest",
            "required": true,
            "schema": {
              "title": "Target Digest",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Target Manifest"
      }
    },
    "/v1/blobs/{digest}": {
      "put": {
        "operationId": "put_blob_v1_blobs__digest__put",
        "parameters": [
          {
            "in": "path",
            "name": "digest",
            "required": true,
            "schema": {
              "title": "Digest",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "content-type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content-Type"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Blob"
      }
    },
    "/v1/extensions": {
      "get": {
        "operationId": "list_extensions_v1_extensions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "title": "Response List Extensions V1 Extensions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Extensions"
      }
    },
    "/v1/extensions/{namespace}/{name}": {
      "get": {
        "operationId": "get_extension_v1_extensions__namespace___name__get",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Extension V1 Extensions  Namespace   Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Extension"
      }
    },
    "/v1/extensions/{namespace}/{name}/versions/{version}": {
      "get": {
        "operationId": "get_release_v1_extensions__namespace___name__versions__version__get",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Release"
      }
    },
    "/v1/extensions/{namespace}/{name}/versions/{version}/publish": {
      "post": {
        "operationId": "publish_release_v1_extensions__namespace___name__versions__version__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Release"
      }
    },
    "/v1/extensions/{namespace}/{name}/versions/{version}/targets/{target_key}": {
      "put": {
        "operationId": "put_target_v1_extensions__namespace___name__versions__version__targets__target_key__put",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "target_key",
            "required": true,
            "schema": {
              "title": "Target Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetAssociation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Target"
      }
    },
    "/v1/extensions/{namespace}/{name}/versions/{version}/yank": {
      "post": {
        "operationId": "yank_release_v1_extensions__namespace___name__versions__version__yank_post",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Yank Release"
      }
    }
  }
}